Working MySQL DB Example
I had to rearrange some stuff in the project, or rather rearranging
happened due to re-orienting myself to a Java project. If this can or
should be arranged better, feel free to do so. As of this commit we have
a simple hello world app that also connects to the db and reads
character names to the console.