Application can now connect to a remote MySQL server.
Adjustments were made in the login to specify a server running in a
location other than localhost. Connection String required sertimezone to
be set to work. A remote server has been said up on my private hosting.
The default values in the login will connect to that server. You can
still enter your own local server values and connect to that as well.

The notable difference is that the db name is now a field to fill out in
server settings. This is because the DB name may not be the same
everywhere it is used.