Player/Admin ListView now show elements in Player ArrayList
Stuff pulled from the DB can be displayed in the view! Also, the DAL are
generally great, but I had to make a minor(ish) change as I go. Instead
of each DAL using a new instance of the connection class, they're passed
the one created at Login. This is because there is a lot of dynamic info
pulled for the connection string, and I didn't want to pass those
variable all over the place to every new instance of a connection.
It'snot a big deal, everything is generally in good order. There is
probably a better way to do this by making our connection class static
(singleton?)