Admin status now updates correctly when giving or taking it from Player
is_active field added to Admin Table. When a user is created, if it is
an admin and doesn't have an entry in the Admin table, one is created
and is_active is set to true. If an entry for the Player ID is found to
exist, its is_active field is updated to the value of the CheckBox on
the edit Player/Admin view.