Renamed Items model to Item for consistency
It is a model for a single Item - so that made sense I think. I put the
bare minimum in for the ItemDAL. That class did not yet exist and I
needed it to at least exist to continue patterning the other iteration 1
configuration options after players/admins. Pushing now, because I'm
tired and making mistakes. This is the last spot for my days work that
makes sense.