New Items can now be created
With this commit, all of our functional goals for Iteration 1 appear to
have been reached. We should still look over the code before calling it.
Some tests no longer pass. It appears some tests are against the DB that
is being used in the actual program, this means the tests will fail if a
user changes the data. We may want to look at a dedicated test db that
we don't actually connect to with the program, just for the hard coded
tests.