Refactored the Abstract and Concrete Quest Control to one control
My effort to simplify with sub classing actually made things more
complex. While there are different templates for entering quests, the
templates are only the views, or how the user reads the data. The data
itself is all input the same way in the database. That is the database
doesn't know about the templates; so why am I making a bunch of
different controller implementations for it?