Merged with latest and conflicts resolved. See description.
Specifically in the QuestContoller and Quest DAL there were a couple of
changes I needed to make to do what needed to be done with the data. It
was pretty minor (changed return type of create quest to int so that it
returns the id of the new quest) and I added get quests by conflict id.

There is still an issue with saving quest edits back to the db. I'll
pick up there tomorrow.