save items for quest now display when editing, commented QuestItemsDAL
QuestItemsDAL was made in previous commit, but forgot to mention. It was
necessary to complete the Quest editing. At this time Quest editing
appear to be fully functional for Iteration 2 expectations. Moving on to
Conflict.

NOTE: it appears some thinking was transposed on the relationship between quest and conflict. A conflict should reference many quests, and a quest can reference one conflict. Using the CD catalog assignment as an example, a conflict is like a cd, and quests are tracks on the cd. To that extent, we actually built this backward, but that is okay. Quests all relate to a hard coded single conflict. There were some db changes needed. Modified db scripts will be in next commit.