Exciting commit where the conflict - quest - quest items all work!
Functionality for Iteration 2 of the Quest Management system has been
realized! The next thing is to replicate "The Quest" conflict template
for the other 3 template types. Before doing that I'll probably take a
moment to tidy up existing code, and perhaps name things better before
replicating.

Beyond this there is just the game story manager. That is all but done.
It may not be implemented until Monday morning.