Updated DB Creation script. Updated Default server info for remote
Default login for server is correct as of 7.6.2020. This will likely be
invalidated in the future.

Collation note. If you run into any errors regarding colation on the DB
script, see https://www.mediawiki.org/wiki/Topic:Szjssrh2k9y431pm for
solution. I had to manually update the script for it to work on the
remote server.