Rolled back work to consolidate create view methods
As it stands, Java is strictly pass-by-value and not pass-by-reference.
This means that I cannot pass in a private field and have that private
field be changed by the parameter name that "references" it. As soon as
I set the passed in View Control, it becomes a different object. There
is likely a way to go about this, but since what exists isn't bad code
and it works, I may just leave it until a time when I can dive a little
deeper in accomplishing the equivilent of pass by reference in Java.