Added player level cap to DB and as editable item in GameStory Manager
the int cell player_character_level_cap must be added to the database
for this to continue working. Also, I added npc_character_level_cap to
the db, same as for player. One can either manually update that
gamestory table - or wait for a new db creation script to be committed.