NPC character position and level implemented in create and edit views
character_level needed to be added to the npccharacters table as an int.
You'll need to add this to your local db, or wait for an updated db
creation script. The online db will reflect this shortly; also, the
npccharacter model needed to be updated for character level filed, so it
will need a test written.