Updated code to reflect Java standard casing practices
Previously the use of upper case first letters was applied according to
how it is done in a Unity C# project. I set that standard early on and
so all of our code was not following Java's practice of using capitol
first letters only for things like class declarations, constructors and
final fields.