Individual Quest Edit View now shows on mainstage, not a new stage
Everything now takes place in one windowed view. Prior to this editing a
quest  opened a separate program window that overlaid the first. These
to windows gaining and/or loosing focus may be have been the cause of
some user issues where things didn't appear to update. Overall, this is
better for the user experience.