LoginStage and MainStage are no longer resizeable
This also means they can only be minimized, not maximized. At this time,
the program is designed around an 800x600 viewing standard. It provides
enough room to see and edit values. If in the future it seems usefully
to make the program resizeable, then the resizing of all elements will
be addressed. As it stands, it looks and works good at its intended
dimensions.